Significant refactoring
Changed the name of the main class to NanoTimer. Also added named splits
and modified the toString method to use them (along with an additional
toString method to specify the TimeUnit for the elapsed time). Created
actual tests from the scenarios that were originally in the App object.
